Separando backend e configurando Docker para rodar corretamente o frontend

+++ .docker/entrypoint.sh
@@ -1,8 +1,19 @@
 #!/bin/bash
 
 #On error no such file entrypoint.sh, execute in terminal - dos2unix .docker\entrypoint.sh
-cp .env.example .env
-cp .env.testing.example .env.testing
+###FRONT_END
+
+#npm config set cache /var/www/.npm-cache --global
+cd /var/www/frontend && npm install && cd ..
+
+###FRONT_END
+cd backend
+if [ ! -f ".env" ]; then
+    cp .env.example .env
+fi
+if [ ! -f ".env.testing" ]; then
+    cp .env.testing.example .env.testing
+fi
 chown -R www-data:www-data .
 composer install
 php artisan key:generate
